+++ miden-lib/src/accounts/wallets/mod.rs
@@ -5,7 +5,7 @@ use alloc::{
 
 use miden_objects::{
     accounts::{
-        Account, AccountCode, AccountId, AccountStorage, AccountStorageType, AccountType, SlotItem,
+        Account, AccountCode, AccountId, AccountStorage, AccountStorageMode, AccountType, SlotItem,
     },
     AccountError, Word,
 };
@@ -30,7 +30,7 @@ pub fn create_basic_wallet(
     init_seed: [u8; 32],
     auth_scheme: AuthScheme,
     account_type: AccountType,
-    account_storage_type: AccountStorageType,
+    account_storage_mode: AccountStorageMode,
 ) -> Result<(Account, Word), AccountError> {
     if matches!(account_type, AccountType::FungibleFaucet | AccountType::NonFungibleFaucet) {
         return Err(AccountError::AccountIdInvalidFieldElement(
@@ -59,7 +59,7 @@ pub fn create_basic_wallet(
     let account_seed = AccountId::get_account_seed(
         init_seed,
         account_type,
-        account_storage_type,
+        account_storage_mode,
         account_code.commitment(),
         account_storage.root(),
     )?;
@@ -76,7 +76,7 @@ mod tests {
     use miden_objects::{crypto::dsa::rpo_falcon512, ONE};
     use vm_processor::utils::{Deserializable, Serializable};
 
-    use super::{create_basic_wallet, Account, AccountStorageType, AccountType, AuthScheme};
+    use super::{create_basic_wallet, Account, AccountStorageMode, AccountType, AuthScheme};
 
     #[test]
     fn test_create_basic_wallet() {
@@ -85,7 +85,7 @@ mod tests {
             [1; 32],
             AuthScheme::RpoFalcon512 { pub_key },
             AccountType::RegularAccountImmutableCode,
-            AccountStorageType::OnChain,
+            AccountStorageMode::Public,
         );
 
         wallet.unwrap_or_else(|err| {
@@ -100,7 +100,7 @@ mod tests {
             [1; 32],
             AuthScheme::RpoFalcon512 { pub_key },
             AccountType::RegularAccountImmutableCode,
-            AccountStorageType::OnChain,
+            AccountStorageMode::Public,
         )
         .unwrap()
         .0;