+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.6.0 (TBD)
 
 - Implemented offset based storage access (#843).
+- [BREAKING] `AccountStorageType` enum was renamed to `AccountStorageMode` along with its variants (#854).
 - [BREAKING] `AccountStub` structure was renamed to `AccountHeader` (#855).
 
 ## 0.5.1 (2024-08-28) - `miden-objects` crate only